Improve function calling samples
